Move parameter lookup for tokens into GitlabWebhookSubscriber
Refs https://github.com/iwink/gitlab-webhook-bundle/issues/10
