--- src/Annotation/Reflection/ReflectionProperty.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Iwink\GitLabWebhookBundle\Annotation\Reflection;
-
-/**
- * Parameters can be resolved from doc blocks.
- * @since 1.1.0
- */
-class ReflectionProperty extends \ReflectionProperty {
-	use ParameterAwareTrait;
-}
